move fire mask compositor into composites init
